+++ Cargo.toml
@@ -148,6 +148,7 @@ semver = "1.0.5"
 serde = { version = "1.0.192", features = [ "derive" ] }
 serde_json = { version = "1.0", features = [ "arbitrary_precision" ] }
 serde_with = "2.3.1"
+similar-asserts = "1.5.0"
 smol_str = { version = "0.2.0", features = [ "serde" ] }
 sqlx = { version = "0.7.2", features = [ "chrono", "macros", "regexp", "runtime-async-std", "runtime-tokio", "sqlite", "uuid" ] }
 starknet = "0.8.0"
