+++ crates/katana/executor/Cargo.toml
@@ -26,3 +26,5 @@ tokio.workspace = true
 cairo-vm.workspace = true
 katana-provider.workspace = true
 rstest.workspace = true
+serde_json.workspace = true
+similar-asserts.workspace = true
