+++ crates/katana/executor/tests/fixtures/contract.json
@@ -1,0 +1,1 @@
+../../../primitives/contracts/compiled/oz_account_080.json
