+++ crates/katana/executor/tests/fixtures/legacy_contract.json
