+++ crates/katana/executor/tests/simulate.rs
@@ -1,0 +1,1 @@
+mod fixtures;
